requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyMySQL==1.0.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
SQLAlchemy==1.4.4
urllib3==1.26.4
Werkzeug==1.0.1
virtualenv
honeycomb-beeline
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-wsgi
opentelemetry-exporter-otlp-proto-grpc